requirements.txt
robotframework>=3.1.0
robotlibcore-temp
